+++ src/main/java/au/com/fx/converter/service/handler/RateConversionHandlerImpl.java
@@ -4,6 +4,8 @@ import au.com.fx.converter.domain.ConversionChart;
 import au.com.fx.converter.service.factory.RateConversionHandlerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.math.BigDecimal;
+
 /**
  * This class will be responsible for finding the appropriate handler and delegate the Rate Conversion.
  * This class will use the {@link RateConversionHandlerFactory} to obtain the Delegate.
@@ -19,7 +21,7 @@ public class RateConversionHandlerImpl implements RateConversionHandler {
     RateConversionHandlerFactory rateConversionHandlerFactory;
 
     @Override
-    public Double process(ConversionChart chart, Double currentRate) {
+    public BigDecimal process(ConversionChart chart, BigDecimal currentRate) {
         if (chart == null)
             return currentRate;
